Rebase and fix release pipeline name
